allow multi-select option, working add/remove/edit, new materials list
