+++ client/js/filament-manager.runner.js
@@ -1,799 +1,44 @@
-import UI from "./utils/ui.js";
-import {setupFilamentManagerReSyncBtn} from "./services/filament-manager-plugin.service";
-import * as ApexCharts from "apexcharts";
-import OctoFarmClient from "./services/octofarm-client.service";
-import {dashboardOptions} from "./pages/charts/dashboard.options";
+import {setupFilamentManagerReSyncBtn} from "./services/octoprint/filament-manager-plugin.service";
+import {
+  loadPageStatistics,
+  renderFilamentUsageCharts, renderMaterialsList, renderProfilesManagerTable,
+  renderSpoolsManagerTable,
+  setupMaterialsPreSelect,
+  setupAddSpoolsListener,
+  setupAddProfilesListener, addProfileTableListeners, addSpoolTableListeners
+} from "./pages/filament-manager/filament-manager.utils";
+import {  updateProfileDrop, updatePrinterDrops } from "./pages/filament-manager/filament-manager-ui.utils";
 
-const pageElements = {
-  filamentProfileTotals: document.getElementById("filamentProfileTotals"),
-  filamentSpoolTotals: document.getElementById("filamentSpoolsTotals"),
-  filamentSpoolsActiveTotals: document.getElementById("filamentSpoolsActiveTotals"),
-  filamentUsedProgress: document.getElementById("filamentUsedProgress"),
-  filamentRemainingProgress: document.getElementById("filamentRemainingProgress"),
-  filamentOverviewTable: document.getElementById("filamentOverviewTable"),
-  filamentOverviewTableHeader: document.getElementById("filamentOverviewTableHeader"),
-  filamentOverviewList: document.getElementById("filamentOverviewList")
-};
 
-let filamentManagerSettings = false;
 
-const filamentStore = [
-  {
-    code: "pla",
-    display: "PLA",
-    density: "1.24"
-  },
-  {
-    code: "abs",
-    display: "ABS",
-    density: "1.04"
-  },
-  {
-    code: "petg",
-    display: "PETG",
-    density: "1.27"
-  },
-  {
-    code: "nylon",
-    display: "NYLON",
-    density: "1.52"
-  },
-  {
-    code: "tpu",
-    display: "TPU",
-    density: "1.21"
-  },
-  {
-    code: "pc",
-    display: "Polycarbonate (PC)",
-    density: "1.3"
-  },
-  {
-    code: "wood",
-    display: "Wood Fill",
-    density: "1.28"
-  },
-  {
-    code: "carbon",
-    display: "Carbon Fibre",
-    density: "1.3"
-  },
-  {
-    code: "pcabs",
-    display: "PC/ABS",
-    density: "1.19"
-  },
-  {
-    code: "hips",
-    display: "HIPS",
-    density: "1.03"
-  },
-  {
-    code: "pva",
-    display: "PVA",
-    density: "1.23"
-  },
-  {
-    code: "asa",
-    display: "ASA",
-    density: "1.05"
-  },
-  {
-    code: "pp",
-    display: "Polypropylene (PP)",
-    density: "0.9"
-  },
-  {
-    code: "acetal",
-    display: "Acetal (POM)",
-    density: "1.4"
-  },
-  {
-    code: "pmma",
-    display: "PMMA",
-    density: "1.18"
-  },
-  {
-    code: "fpe",
-    display: "Semi Flexible FPE",
-    density: "2.16"
-  }
-];
 
-async function reRenderPageInformation() {
-  const { statistics } = await OctoFarmClient.getFilamentStatistics();
-  const { profiles } = await OctoFarmClient.getFilamentProfiles();
-  const { spools } = await OctoFarmClient.getFilamentSpools();
-  pageElements.filamentProfileTotals.innerHTML = statistics.profileCount;
-  pageElements.filamentSpoolTotals.innerHTML = statistics.spoolCount;
-  pageElements.filamentSpoolsActiveTotals.innerHTML = statistics.activeSpoolCount;
-  const usedPercent = ((statistics.used / statistics.total) * 100).toFixed(0);
-  pageElements.filamentUsedProgress.innerHTML = usedPercent + "%";
-  pageElements.filamentUsedProgress.style.width = usedPercent;
-  const remainingPercent = (
-    ((statistics.total - statistics.used) / statistics.total) *
-    100
-  ).toFixed(0);
-  pageElements.filamentRemainingProgress.innerHTML = remainingPercent + "%";
-  pageElements.filamentRemainingProgress.style.width = remainingPercent;
-
-  let headerBreakdown = "";
-  statistics.materialBreakDown.forEach((used) => {
-    headerBreakdown += `<th scope="col">${used.name}</th>`;
-  });
-
-  let remainingBreakdown = "";
-  statistics.materialBreakDown.forEach((used) => {
-    remainingBreakdown += `<th scope="col">${(used.total - used.used).toFixed(2) / 1000}kg</th>`;
-  });
-  let usedBreakdown = "";
-  statistics.materialBreakDown.forEach((used) => {
-    usedBreakdown += `<th scope="col">${used.used.toFixed(2) / 1000}kg</th>`;
-  });
-
-  let weightBreakdown = "";
-  statistics.materialBreakDown.forEach((used) => {
-    weightBreakdown += `<th scope="col">${used.total.toFixed(2) / 1000}kg</th>`;
-  });
-
-  let costBreakdown = "";
-  statistics.materialBreakDown.forEach((used) => {
-    costBreakdown += `<th scope="col">${used.price.toFixed(2)}</th>`;
-  });
-
-  pageElements.filamentOverviewTableHeader.innerHTML = `
-  <tr>
-    <td scope="row">Material Overview: </td>
-    ${headerBreakdown}    
-  </tr>
-  `;
-
-  pageElements.filamentOverviewTable.innerHTML = `
-  <tr>
-      <td scope="row">Remaining <span class="badge badge-success ml-2">${
-        (statistics.total.toFixed(0) - statistics.used.toFixed(0)) / 1000
-      }kg</span></td>
-      ${remainingBreakdown}
-  </tr>
-  <tr>
-      <td scope="row">Used <span class="badge badge-warning ml-2">${
-        statistics.used.toFixed(0) / 1000
-      }kg</span> </td>
-      ${usedBreakdown}
-  </tr>
-  <tr>
-      <th scope="row">Weight<span class="badge badge-light text-dark ml-2">${
-        statistics.total.toFixed(0) / 1000
-      }kg</span></th>
-      ${weightBreakdown}
-  </tr>
-  <tr>
-      <th scope="row">Cost <span class="badge badge-info ml-2">${statistics.price.toFixed(
-        2
-      )}</span></th>
-      ${costBreakdown}
-  </tr>
-  `;
-
-  let spoolList = "";
-  spools.forEach((spool) => {
-    spoolList += `
-                <tr>
-                WHERE IS THIS
-                    <th style="display: none;"> </th>
-                    <th><i class="fas fa-toilet-paper"></i></th>
-                    <th scope="row">${spool.name}</th>
-                    <td id="spoolsListMaterial-${spool._id}">
-                    </td>
-                    <td id="spoolsListManufacture-${spool._id}">
-                    </td>
-                    <td>${spool.weight / 1000}KG</td>
-                    <td>${(spool.weight - spool.used).toFixed(0)}g</td>
-                    <td>${spool.tempOffset}</td>
-                    <td>${spool.bedOffset || 0}</td>
-                    <td id="spoolsListPrinterAssignment-${spool._id}"></td>
-                </tr>
-    `;
-  });
-
-  pageElements.filamentOverviewList.innerHTML = spoolList;
-
-  jpList.refresh()
-}
-
-// Profile functions
-async function addProfile(manufacturer, material, density, diameter) {
-  const errors = [];
-
-  if (manufacturer.value === "") {
-    errors.push({ type: "warning", msg: "Please input manufacturer" });
-  }
-  if (material.value === "") {
-    errors.push({ type: "warning", msg: "Please select or type a material" });
-  }
-  if (density.value === 0) {
-    errors.push({ type: "warning", msg: "Please input a density" });
-  }
-  if (diameter.value === 0) {
-    errors.push({ type: "warning", msg: "Please input a density" });
-  }
-  if (errors.length > 0) {
-    errors.forEach((error) => {
-      UI.createMessage(error, "profilesMessage");
-    });
-    return;
-  }
-  const opts = {
-    manufacturer: manufacturer.value,
-    material: material.value,
-    density: density.value,
-    diameter: diameter.value
-  };
-  let post = await OctoFarmClient.post("filament/save/profile", opts);
-  if (post && post.errors.length === 0) {
-    UI.createMessage(
-      {
-        type: "success",
-        msg: "Successfully added new profile to the database..."
-      },
-      "profilesMessage"
-    );
-    manufacturer.value = "";
-    material.value = "";
-    density.value = 1.25;
-    diameter.value = 1.75;
-    await reRenderPageInformation();
-    await updateProfileDrop();
-  } else {
-    printErrors(post.errors)
-  }
-}
-
-function printErrors(errors) {
-  let errorMessage = "OctoFarm is having issues! Error(s): <br>";
-  errors.forEach(error => {
-    errorMessage += error + "<br>"
-  })
-  UI.createAlert("error", errorMessage, 6000, "Clicked");
-}
-
-async function editProfile(e) {
-  const row = e.parentElement.parentElement;
-  const editable = row.querySelectorAll("input");
-  const id = e.parentElement.parentElement.firstElementChild.innerHTML.trim();
-  editable.forEach((edit) => {
-    edit.disabled = false;
-    edit.value = edit.placeholder;
-  });
-  document.getElementById(`save-${id}`).classList.remove("d-none");
-  document.getElementById(`edit-${id}`).classList.add("d-none");
-}
-async function saveProfile(e) {
-  const row = e.parentElement.parentElement;
-  const editable = row.querySelectorAll("input");
-  const id = e.parentElement.parentElement.firstElementChild.innerHTML.trim();
-  const profile = [];
-  editable.forEach((edit) => {
-    edit.disabled = true;
-    edit.placeholder = edit.value;
-    profile.push(edit.value);
-    edit.value = "";
-  });
-  const data = {
-    id,
-    profile
-  };
-  let post = await OctoFarmClient.post("filament/edit/profile", data);
-  if (post && post.errors.length === 0) {
-    await reRenderPageInformation();
-    await updateProfileDrop();
-    document.getElementById(`save-${id}`).classList.add("d-none");
-    document.getElementById(`edit-${id}`).classList.remove("d-none");
-  }else{
-    printErrors(post.errors)
-  }
-  jplist.refresh();
-}
-async function deleteProfile(e) {
-  document.getElementById("profilesMessage").innerHTML = "";
-  if (e.classList.contains("delete") || e.classList.contains("deleteIcon")) {
-    let post = await OctoFarmClient.post("filament/delete/profile", {
-      id: e.parentElement.parentElement.firstElementChild.innerHTML.trim()
-    });
-    if (post && post.errors.length === 0) {
-      if (e.classList.contains("deleteIcon")) {
-        jplist.resetContent(function () {
-          // remove element with id = el1
-          e.parentElement.parentElement.parentElement.remove();
-        });
-      } else {
-        jplist.resetContent(function () {
-          // remove element with id = el1
-          e.parentElement.parentElement.remove();
-        });
-      }
-      await reRenderPageInformation();
-      await updateProfileDrop();
-      UI.createAlert("success", "Successfully deleted your profile!", 3000, "Clicked")
-    } else {
-      printErrors(post.errors)
-    }
-  }
-}
-
-// Spool functions
-async function addSpool(
-  spoolsName,
-  spoolsProfile,
-  spoolsPrice,
-  spoolsWeight,
-  spoolsUsed,
-  spoolsTempOffset,
-  spoolsBedOffset
-) {
-  const errors = [];
-
-  if (spoolsName.value === "") {
-    errors.push({ type: "warning", msg: "Please input a spool name" });
-  }
-  if (spoolsProfile.value === "") {
-    errors.push({ type: "warning", msg: "Please select a profile" });
-  }
-  if (spoolsPrice.value === "") {
-    errors.push({ type: "warning", msg: "Please input a spool price" });
-  }
-  if (spoolsWeight.value === 0 || spoolsWeight.value === "") {
-    errors.push({ type: "warning", msg: "Please input a spool weight" });
-  }
-  if (spoolsUsed.value === "") {
-    errors.push({ type: "warning", msg: "Please input spool used weight" });
-  }
-
-  if (errors.length > 0) {
-    errors.forEach((error) => {
-      UI.createMessage(error, "addSpoolsMessage");
-    });
-    return;
-  }
-  const opts = {
-    spoolsName: spoolsName.value,
-    spoolsProfile: spoolsProfile.value,
-    spoolsPrice: spoolsPrice.value,
-    spoolsWeight: spoolsWeight.value,
-    spoolsUsed: spoolsUsed.value,
-    spoolsTempOffset: spoolsTempOffset.value,
-    spoolsBedOffset: spoolsBedOffset.value
-  };
-  let post = await OctoFarmClient.post("filament/save/filament", opts);
-
-  if (post || post?.errors?.length === 0) {
-    UI.createMessage(
-      {
-        type: "success",
-        msg: "Successfully added new roll to the database..."
-      },
-      "addSpoolsMessage"
-    );
-
-    spoolsName.value = "";
-    spoolsPrice.value = "";
-    spoolsWeight.value = 1000;
-    spoolsUsed.value = 0;
-    spoolsTempOffset.value = 0.0;
-    spoolsBedOffset.value = 0;
-    await reRenderPageInformation();
-    await updatePrinterDrops();
-    await updateProfileDrop();
-    return true;
-  } else {
-    printErrors(post.errors)
-  }
-}
-
-const clonedSpools = [];
-
-async function cloneSpool(e) {
-  const row = e.parentElement.parentElement;
-  const editable = row.querySelectorAll("input");
-  const selects = row.querySelectorAll("select");
-  const spool = [];
-  editable.forEach((edit) => {
-    spool.push(edit.placeholder);
-  });
-
-  let clonedIndex = clonedSpools.length;
-
-  document.getElementById("addSpoolsTable").insertAdjacentHTML(
-    "afterend",
-    `
-              <tr id="clonedRow-${clonedIndex}">
-                <th style="display: none;">${clonedIndex}</th>
-                <th scope="row"><input id="clonedName-${clonedIndex}" class="form-control" type="text" placeholder="${spool[0]}-${clonedIndex}" value="${spool[0]} (${clonedIndex})"></th>
-                <td>
-                     <span class="d-none material" id="spoolsMaterialText-${clonedIndex}"></span>
-                     <select id="spoolsProfile-${clonedIndex}" class="form-control">
-
-                     </select>
-                 </td>
-                <td><input id="clonedPrice-${clonedIndex}" class="form-control" type="text" step="0.01" placeholder="${spool[1]}" value="${spool[1]}"></td>
-                <td><input id="clonedWeight-${clonedIndex}" class="form-control" type="text" placeholder="${spool[2]}" value="1000"></td>
-                <td><input id="clonedUsed-${clonedIndex}" class="form-control" type="text" placeholder="${spool[3]}" value="0"></td>
-                <td><input id="clonedToolOffset-${clonedIndex}" class="form-control" type="text" placeholder="${spool[4]}" value="${spool[4]}"></td>
-                <td><input id="clonedBedOffset-${clonedIndex}" class="form-control" type="text" placeholder="${spool[5]}" value="${spool[5]}"></td>
-                 <td>
-                     <select id="spoolsPrinterAssignment-${clonedIndex}" class="form-control" disabled>
-
-                     </select>
-                 </td>
-                 <td>
-                <button id="clonedSave-${clonedIndex}" type="button" class="btn btn-sm btn-success">
-                  <i class="fas fa-save saveIcon"></i>
-                </button></td>
-              </tr>
-              `
-  );
-
-  document.getElementById("clonedSave-" + clonedIndex).addEventListener("click", function () {
-    const clonedRow = document.getElementById(`clonedRow-${clonedIndex}`);
-    const addedSpool = addSpool(
-      document.getElementById(`clonedName-${clonedIndex}`),
-      document.getElementById(`spoolsProfile-${clonedIndex}`),
-      document.getElementById(`clonedPrice-${clonedIndex}`),
-      document.getElementById(`clonedWeight-${clonedIndex}`),
-      document.getElementById(`clonedUsed-${clonedIndex}`),
-      document.getElementById(`clonedToolOffset-${clonedIndex}`),
-      document.getElementById(`clonedBedOffset-${clonedIndex}`)
-    );
-    if (addedSpool) {
-      // Remove Row
-      clonedRow.remove();
-      clonedSpools.pop();
-    }
-  });
-
-  clonedSpools.push(clonedIndex);
-  document.getElementById(`spoolsProfile-${clonedIndex}`).value = selects[0].value;
-  await updateProfileDrop();
-}
-
-async function editSpool(e) {
-  const row = e.parentElement.parentElement;
-  const editable = row.querySelectorAll("input");
-  const id = e.parentElement.parentElement.firstElementChild.innerHTML.trim();
-  editable.forEach((edit) => {
-    edit.disabled = false;
-    edit.value = edit.placeholder;
-  });
-  document.getElementById(`spoolsPrinterAssignment-${id}`).disabled = false;
-  document.getElementById(`spoolsProfile-${id}`).disabled = false;
-  document.getElementById(`save-${id}`).classList.remove("d-none");
-  document.getElementById(`edit-${id}`).classList.add("d-none");
-  await reRenderPageInformation();
-}
-async function deleteSpool(e) {
-  document.getElementById("profilesMessage").innerHTML = "";
-  if (e.classList.contains("delete") || e.classList.contains("deleteIcon")) {
-    let post = await OctoFarmClient.post("filament/delete/filament", {
-      id: e.parentElement.parentElement.firstElementChild.innerHTML.trim()
-    });
-    if (post?.spool) {
-      await reRenderPageInformation();
-      if (e.classList.contains("deleteIcon")) {
-        jplist.resetContent(function () {
-          // remove element with id = el1
-          e.parentElement.parentElement.parentElement.remove();
-        });
-      } else {
-        jplist.resetContent(function () {
-          // remove element with id = el1
-          e.parentElement.parentElement.remove();
-        });
-      }
-    } else {
-      UI.createAlert(
-        "error",
-        "There was a conflict when deleting this spool, is it attached to a printer?",
-        4000,
-        "Clicked"
-      );
-    }
-  }
-}
-async function saveSpool(e) {
-  const row = e.parentElement.parentElement;
-  const editable = row.querySelectorAll("input");
-  const id = e.parentElement.parentElement.firstElementChild.innerHTML.trim();
-  const spool = [];
-  editable.forEach((edit) => {
-    edit.disabled = true;
-    edit.placeholder = edit.value;
-    spool.push(edit.value);
-    edit.value = "";
-  });
-
-  spool.push(document.getElementById(`spoolsProfile-${id}`).value);
-  const data = {
-    id,
-    spool
-  };
-  let post = await OctoFarmClient.post("filament/edit/filament", data);
-  if (post) {
-    document.getElementById(`spoolsPrinterAssignment-${id}`).disabled = true;
-    document.getElementById(`spoolsProfile-${id}`).disabled = true;
-    document.getElementById(`save-${id}`).classList.add("d-none");
-    document.getElementById(`edit-${id}`).classList.remove("d-none");
-  }
-
-  await selectFilament(
-    getSelectValues(document.getElementById(`spoolsPrinterAssignment-${id}`)),
-    id
-  );
-  jplist.refresh();
-}
-
-async function updateProfileDrop() {
-  // Update filament selection profile drop
-  const spoolsProfile = document.getElementById("spoolsProfile");
-  const { profiles } = await OctoFarmClient.getFilamentProfiles()
-  const  { spools }  = await OctoFarmClient.getFilamentSpools()
-  if (!!profiles) {
-    spoolsProfile.innerHTML = "";
-    profiles.forEach((profile) => {
-      let profileID = null;
-      profileID = profile?._id;
-      spoolsProfile.insertAdjacentHTML(
-        "beforeend",
-        `
-             <option value="${profileID}">${profile?.manufacturer} (${profile?.material})</option>
-            `
-      );
-    });
-  }
-  // Generate profile assignment
-  const printerDrops = document.querySelectorAll("[id^='spoolsProfile-']");
-  const printerListMaterials = document.querySelectorAll("[id^='spoolsListMaterial-']");
-  const spoolsListManufacture = document.querySelectorAll("[id^='spoolsListManufacture-']");
-  const spoolsMaterialText = document.querySelectorAll("[id^='spoolsMaterialText-']");
-  printerDrops.forEach((drop, index) => {
-    drop.innerHTML = "";
-    profiles.forEach((prof) => {
-      drop.insertAdjacentHTML(
-        "beforeend",
-        `<option value="${prof?._id}">${prof?.manufacturer} (${prof?.material})</option>`
-      );
-    });
-    const spoolID = drop?.id.split("-");
-    const spool = _.findIndex(spools, function (o) {
-      return o?._id === spoolID[1];
-    });
-    if (typeof spools[spool] !== "undefined") {
-      drop.value = spools[spool].profile;
-      const profileID = _.findIndex(profiles, function (o) {
-        return o._id === spools[spool].profile;
-      });
-      drop.className = `form-control ${profiles[profileID]?.material.replace(/ /g, "_")}`;
-      spoolsMaterialText[index].innerHTML = `${profiles[profileID]?.material}`;
-    }
-  });
-  //Fix for not updating main spool list with correct information, not skipping fo shizzle
-  spoolsListManufacture.forEach((text) => {
-    const spoolID = text.id.split("-");
-    const spool = _.findIndex(spools, function (o) {
-      return o._id === spoolID[1];
-    });
-    if (typeof spools[spool] !== "undefined") {
-      const profileID = _.findIndex(profiles, function (o) {
-        return o._id === spools[spool]?.profile;
-      });
-      text.innerHTML = `${profiles[profileID]?.manufacturer}`;
-    }
-  });
-  printerListMaterials.forEach((text) => {
-    const spoolID = text.id.split("-");
-    const spool = _.findIndex(spools, function (o) {
-      return o._id === spoolID[1];
-    });
-    if (typeof spools[spool] !== "undefined") {
-      const profileID = _.findIndex(profiles?.profiles, function (o) {
-        return o._id === spools[spool]?.profile;
-      });
-      text.innerHTML = `${profiles[profileID]?.material}`;
-    }
-  });
-}
-
-function getSelectValues(select) {
-  const result = [];
-  const options = select && select.options;
-  let opt;
-
-  for (let i = 0, iLen = options.length; i < iLen; i++) {
-    opt = options[i];
-
-    if (opt.selected) {
-      result.push(opt.value || opt.text);
-    }
-  }
-  return result;
-}
-
-async function updatePrinterDrops() {
-  let filament = await OctoFarmClient.get("filament/get/filament");
-  let { printerList } = await OctoFarmClient.get("filament/get/printerList");
-  let printerDropList = "";
-  printerList.forEach((printer) => {
-    printerDropList += printer;
-  });
-
-  const printerDrops = document.querySelectorAll("[id^='spoolsPrinterAssignment-']");
-  const printerAssignments = document.querySelectorAll("[id^='spoolsListPrinterAssignment-']");
-  printerDrops.forEach((drop, index) => {
-    const split = drop.id.split("-");
-    const spoolID = split[1];
-    const spool = _.findIndex(filament?.Spool, function (o) {
-      return o?._id == spoolID;
-    });
-    drop.innerHTML = printerDropList;
-
-    const selectedValues = [];
-
-    if (typeof filament[spool] !== "undefined") {
-      filament[spool]?.printerAssignment.forEach((printer) => {
-        selectedValues.push(`${printer?.id}-${printer?.tool}`);
-      });
-    }
-    for (let i = 0; i < drop.options.length; i++) {
-      if (selectedValues.includes(drop.options[i].value)) {
-        drop.options[i].selected = true;
-      }
-    }
-    drop.disabled = true;
-  });
-  printerAssignments.forEach((text, index) => {
-    text.innerHTML = "";
-    const split = text.id.split("-");
-    const spoolID = split[1];
-    const spoolIndex = _.findIndex(spool, function (o) {
-      return o?._id == spoolID;
-    });
-    if (typeof filament[spoolIndex] !== "undefined") {
-      if (filament[spoolIndex]?.printerAssignment.length > 0) {
-        filament[spoolIndex]?.printerAssignment.forEach((printer) => {
-          text.innerHTML +=
-            "<small>" + printer.name + ": Tool" + printer.tool + "<br>" + "</small>";
-        });
-      } else {
-        text.innerHTML = "Not Assigned";
-      }
-    }
-  });
-}
-
-async function selectFilament(spools, id) {
-  let printerIds = [];
-  for (let i = 0; i < spools.length; i++) {
-    const meta = spools[i].split("-");
-    const printerId = meta[0];
-    printerIds.push({
-      printer: printerId,
-      tool: meta[1]
-    });
-  }
-
-  await OctoFarmClient.post("/filament/assign", {
-    printers: printerIds,
-    spoolId: id
-  });
-
-  await reRenderPageInformation();
-  await updatePrinterDrops();
-}
 
 async function init() {
-  let historyStatistics = await OctoFarmClient.getHistoryStatistics();
-  let usageByDay = historyStatistics.history.totalByDay;
-  let usageOverTime = historyStatistics.history.usageOverTime;
+  await renderFilamentUsageCharts();
 
-  for (let i = 0; i < usageByDay.length; i++) {}
+  await loadPageStatistics();
 
-  if (typeof usageOverTime[0] !== "undefined") {
-    let systemFarmTemp = new ApexCharts(
-      document.querySelector("#usageOverTime"),
-      dashboardOptions.filamentUsageByDayChartOptions
-    );
-    await systemFarmTemp.render();
-    let usageOverFilamentTime = new ApexCharts(
-      document.querySelector("#usageOverFilamentTime"),
-      dashboardOptions.filamentUsageOverTimeChartOptions
-    );
-    await usageOverFilamentTime.render();
+  await renderProfilesManagerTable();
 
-    usageOverFilamentTime.updateSeries(usageOverTime);
-    systemFarmTemp.updateSeries(usageByDay);
-  }
+  addProfileTableListeners();
 
-  // Grab data
-  const spoolTable = document.getElementById("addSpoolsTable");
-  // Initialise materials dropdown
-  const dataList = document.getElementById("profilesMaterial");
-  dataList.addEventListener("change", function (e) {
-    const { value } = this;
-    const selection = _.findIndex(filamentStore, function (o) {
-      return o.code == value.toLowerCase();
-    });
-    if (selection != -1) {
-      this.value = filamentStore[selection].display;
-      document.getElementById("profilesDensity").value = filamentStore[selection].density;
-    }
-  });
-  filamentStore.forEach((filament) => {
-    document.getElementById("huge_list").insertAdjacentHTML(
-      "beforeend",
-      `
-            <option value="${filament.code.toUpperCase()}">${filament.display}</option>
-        `
-    );
-  });
-  spoolTable.addEventListener("click", (e) => {
-    // Remove from UI
-    if (e.target.classList.contains("edit")) {
-      editSpool(e.target);
-    } else if (e.target.classList.contains("delete")) {
-      deleteSpool(e.target);
-    } else if (e.target.classList.contains("save")) {
-      saveSpool(e.target);
-    } else if (e.target.classList.contains("clone")) {
-      cloneSpool(e.target);
-    }
-  });
+  await renderSpoolsManagerTable();
+
+  addSpoolTableListeners();
+
   await updateProfileDrop();
   await updatePrinterDrops();
-  // Initialise Profile Listeners
-  const profilesBtn = document.getElementById("addProfilesBtn");
-  profilesBtn.addEventListener("click", async (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    document.getElementById("profilesMessage").innerHTML = "";
-    const profilesManufactuer = document.getElementById("profilesManufactuer");
-    const profilesMaterial = document.getElementById("profilesMaterial");
-    const profilesDensity = document.getElementById("profilesDensity");
-    const profilesDiameter = document.getElementById("profilesDiameter");
-    await addProfile(profilesManufactuer, profilesMaterial, profilesDensity, profilesDiameter);
-  });
-  document.getElementById("addProfilesTable").addEventListener("click", (e) => {
-    // Remove from UI
-    if (e.target.classList.contains("edit")) {
-      editProfile(e.target);
-    } else if (e.target.classList.contains("delete")) {
-      deleteProfile(e.target);
-    } else if (e.target.classList.contains("save")) {
-      saveProfile(e.target);
-    }
-  });
 
-  //    //Init Spools
-  document.getElementById("addSpoolBtn").addEventListener("click", async (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    document.getElementById("addSpoolsMessage").innerHTML = "";
-    const spoolsName = document.getElementById("spoolsName");
-    const spoolsProfile = document.getElementById("spoolsProfile");
-    const spoolsPrice = document.getElementById("spoolsPrice");
-    const spoolsWeight = document.getElementById("spoolsWeight");
-    const spoolsUsed = document.getElementById("spoolsRemaining");
-    const spoolsTempOffset = document.getElementById("spoolsTempOffset");
-    const spoolsBedOffset = document.getElementById("spoolsBedOffset");
-    await addSpool(
-      spoolsName,
-      spoolsProfile,
-      spoolsPrice,
-      spoolsWeight,
-      spoolsUsed,
-      spoolsTempOffset,
-      spoolsBedOffset
-    );
-  });
   await setupFilamentManagerReSyncBtn();
+
+  renderMaterialsList();
+
+  setupMaterialsPreSelect()
+
+  setupAddSpoolsListener();
+
+  setupAddProfilesListener();
+
 
   jplist.init({
     storage: "localStorage", // 'localStorage', 'sessionStorage' or 'cookies'
@@ -826,15 +71,15 @@ function updateTotals(filtered) {
 const element = document.getElementById("listenerSpools");
 element.addEventListener(
   "jplist.state",
-  (e) => {
+  async (e) => {
     // the elements list after filtering + pagination
     updateTotals(e.jplistState.filtered);
-    updateProfileDrop();
-    updatePrinterDrops();
+    await updateProfileDrop();
+    await updatePrinterDrops();
   },
   false
 );
-init();
+init().then();
 // $("select option").on("dblclick", function (e) {
 //   this.selected = !this.selected;
 //   e.preventDefault();
