+++ server/views/partials/Filament/filamentManager.panel.ejs
@@ -14,6 +14,10 @@
             type="button" class="btn btn-secondary m-2" data-toggle="modal" data-target="#spoolManagerModal">
         <i class="fas fa-toilet-paper"></i> Spool Manager
     </button>
+    <button title="See the current list of materials available to OctoFarm..."
+            type="button" class="btn btn-secondary m-2" data-toggle="modal" data-target="#materialsListModal">
+        <i class="fa-solid fa-brush"></i> Materials List
+    </button>
     <button title="Updates OctoFarms Spools and Profiles database with the filament manager plugin."
             id="resyncFilamentManagerBtn"
             type="button"
